Develop File Modularity
