--- Zap.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Zap</title>
-
-    <link rel="stylesheet" type="text/css" href="Zap.css">
-
-</head>
-<body>
-    
-    <div class="top-bar">
-        <div class="header">
-            <h3>Burger King ▾</h3>
-            <a href="#">☰</a>
-        </div>
-        
-        <div class="course-list">
-            <!-- course links injected by js -->
-        </div>
-    </div>
-
-    <div class="course-sections-body">
-        <!-- course sections injected by js -->
-    </div>
-
-    <script src="Zap.js"></script>
-
-    <div class="bottom-bar">
-
-        <div class="order-total">
-            <p>Total</p>
-            <h2 class="total-price">$14.50</h2>
-            <p class="total-items">6 items</p>
-        </div>
-
-        <a class="order-button"href="#order-screen">order</a>
-
-    </div>
-
-</body>
-</html>
