+++ ZapStylesheets/BarTop.css
@@ -1,0 +1,63 @@
+.top-bar {
+    position: fixed;
+    width: 100%;
+    top: 0;
+    height: 13vh;
+    background-color: var(--primary-background-color);
+    box-shadow: 0 0.5vh 5vh rgb(0, 0, 0, 0.4);
+    z-index: 1000;
+}
+
+.header {
+    text-align: left;
+    padding: 0.0rem 0rem;
+    width: 100%;
+    display: grid;
+    grid-template-columns: 80% 20%;
+    grid-gap: 8%;
+}
+
+.header h3 {
+    height: 1.5rem;
+    margin-left: 5%;
+}
+
+.header a {
+    line-height: 1;
+    color: var(--primary-text-color);
+    text-decoration: none;
+    font-size: 1.3rem;
+    height: 1.5rem;
+    width: min-content;
+}
+
+.course-list {
+    height: auto;
+    padding: 0.4rem 0.6rem;
+    margin: 0.4rem 0;
+    display: flex;
+    gap: 0.4rem;
+    overflow-x: auto;
+    white-space: nowrap;
+    border: 1px solid var(--secondary-color);
+    -ms-overflow-style: none;
+    scrollbar-width: none;
+    border: none;
+}
+
+.course-list::-webkit-scrollbar {
+    display: none;
+}
+
+.course {
+    font-size: 0.8rem;
+    width: min-content;
+    padding: 0.3rem 0.8rem;
+    background-color: var(--primary-button-color);
+    border-radius: 50px;
+}
+
+.course a {
+    color: var(--primary-text-color);
+    text-decoration: none;
+}
