+++ ZapStylesheets/ZapBase.css
@@ -1,0 +1,10 @@
+* {
+    /* border: 1px dotted red; */
+    padding: 0;
+    margin: 0;
+}
+
+html {
+    font-size: 24px;
+    font-family: var(--primary-font);
+}
