+++ ZapStylesheets/ZapBody.css
@@ -1,0 +1,57 @@
+
+.course-sections {
+    padding: 0.6em 0 0 0;
+}
+
+.course-sections-scroller {
+    display: flex;
+    overflow-x: auto;
+    white-space: nowrap;
+}
+
+.course-title {
+    width: max-content;
+    overflow: none;
+    margin-left: 3%;
+}
+
+.course-sections-body {
+    margin: 13vh 0;
+}
+
+.course-section {
+    padding: 0.8em 0;
+    margin-left: 2%;
+    width: min-content;
+    min-width: 200px;
+    max-width: 1000px;
+    border-radius: 25px;
+    height: 40%;
+    font-size: 20px;
+}
+
+.course-text {
+    display: grid;
+    grid-template-columns: 70% 30%;
+    gap: 5%;
+    overflow: none;
+    padding: 0 2.5%;
+}
+
+.course-section h5 {
+    overflow-x: hidden;
+    font-weight: 500;
+    display: flex;
+    align-self: center;
+}
+
+.course-section img {
+    max-width: 460px;
+    width: 46vw;
+    min-width: 200px;
+    max-height: 300px;
+    height: 30vw;
+    min-height: 100px;
+    object-fit: cover;
+    border-radius: 15px;
+}
